+++ kpops/components/streams_bootstrap/streams/streams_app.py
@@ -2,14 +2,13 @@ from __future__ import annotations
 
 from typing import Literal
 
-from pydantic import Extra, Field
+from pydantic import Field
 from typing_extensions import override
 
 from kpops.components.base_components.kafka_app import KafkaApp
 from kpops.components.streams_bootstrap.app_type import AppType
 from kpops.components.streams_bootstrap.streams.model import StreamsAppConfig
 from kpops.utils.docstring import describe_attr, describe_object
-from kpops.utils.pydantic import DescConfig
 
 
 class StreamsApp(KafkaApp):
@@ -38,9 +37,6 @@ class StreamsApp(KafkaApp):
         default=...,
         description=describe_attr("app", __doc__),
     )
-
-    class Config(DescConfig):
-        extra = Extra.allow
 
     @override
     def add_input_topics(self, topics: list[str]) -> None:
